GP tests: don't patch xtest_7000.c, create xtest_7000_gp.c instead
xtest_7000.c contains a subset of the GP tests in TEE.xml (the 29-*
series). When GP tests are enabled, instead of patching xtest_7000.c
we can just generate a new file called xtest_7000_gp.c with all the
tests from TEE.xml. This avoids modifying a source file that is managed
by Git.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>
Tested-by: Pascal Brand <pascal.brand@linaro.org> (STM)