xtest 1008: fix out of memory problem
Fixes out of secure memory problem with running this test case with LPAE.

The out of memory problem was introduced by
38a66f9 Make crypt ta multi-instance

Previously the crypt ta was single-instance, multi-session allowing
several open sessions share the same context (and saving some memory).
When the crypt ta is instead multi-instance test case 1008 indirectly
loads two instances of crypt TA which is too here.

This patch skips the first session to crypt TA which was opened first in
the test case due to legacy reasons.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Tested-by: Jens Wiklander <jens.wiklander@linaro.org> (Juno)
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>