Run xtest 1013 test 50 times
xtest 1013 tests the concurrent TA feature. It lauches several threads
in parallel. Hence, the running behavior is quite sensible, and an
issue in concurrent TA could be raised or not depending on the
runtime.

Running the tests several times allow to run the test in slightly
different configuration, and has a higher probability to raise
potential issues.

Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Pascal Brand <pascal.brand@st.com>