Fix xtest 4007 DH tests
https://github.com/OP-TEE/optee_os/pull/816 breaks "xtest 4007 -l 1".
It turns out some tests are wrong because they provide an incorrect
key_size.

Also, when TEE_ATTR_DH_X_BITS is given, the key generation function is
supposed to use it as the size for the private key to be generated.
xtest was incorrectly checking against key_size.

One test with TEE_ATTR_DH_X_BITS is enabled in the default test suite
(level 0) so that we are more likely to catch any regression.

Depends on https://github.com/OP-TEE/optee_os/pull/816.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>