Avoid writing to the TA directory
Presuming that the tee-supplicant is configured to support loading TAs
from an alternate path, use this alternate path to write test TAs,
instead of trying to rename and overwrite the existing TA images.
This gives several benefits:

- The tests no longer assume that the TA directory is writable.  This
  allows this test to pass on platforms where it is not.

- There is no longer a race with pressing Ctrl-C at an inopportune
  time will leave a corrupted TA file in place instead of the correct
  one.

This change affects test 1008.4 and 1008.5, "Load fake uuid TA", and
"Load corrupt TA".

Signed-off-by: David Brown <david.brown@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey)