Add xtest 6015 (TA storage separation)
xtest 6015 is added to test storage isolation. It invokes a first TA to
create a persistent object. Then it uses a second TA to try and open an
object with the same name, which is expected to fail. Both TAs are
actually the same except for the UUID.

The test also attempts to simulate a file manipulation in the REE file
system (renaming the storage path of TA1 to match the TA2 directory).
This should fail as well.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (QEMU)
Reviewed-by: David Brown <david.brown@linaro.org>
Tested-by: David Brown <david.brown@linaro.org> (Hikey android)