Move GlobalPlatform tests to their own test suite
Introduce a new test suite, selected by "-t gp", and move all the
GlobalPlatform tests there.
In order to maintain some backward compatibility, xtest will run both
the "regression" and "gp" suites in a row by default. To implement this
behavior, accept a "+" as a separator for test suite names, and set the
default test suite name to "regression+gp".

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>