regression_5006: use 32KB buffer
Since OP-TEE client commit 9faeb1b6b023 ("tee_client_api: Correct max
shared memory size define"), TEEC_CONFIG_SHAREDMEM_MAX_SIZE is defined
as ULONG_MAX to signify that there is no fixed limit on the size of
shared memory buffers. This breaks xtest 5006, which assumes the
constant is some "reasonable" size, so that allocation and registration
of such a buffer should succeed.

Let's drop the broken tests. Instead, test allocation and registration
of 32KB, since that was tested in the past and it is good to know our
implementation still supports that (arbitrary) size.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey, QEMU)
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>