Make sure AES GCM does not buffer data
The documentation [1] for TEE_AEUpdate() states that "Unless one or
more calls of this function have supplied sufficient input data, no
output is generated". This can be understood as: "As soon as sufficient
input data have been supplied, some output is generated".

AES GCM effectively works as a stream cipher, so that any amount of
data passed to the algorithm should produce as much output data
immediately. In other words, calling TEE_AEUpdate() with N bytes of
input should generate N bytes of output for any N >= 1.

This commit updates regression_4005 to check the above assertion.

[1] GlobalPlatform TEE Internal Core API Specification v1.1

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>