Add regression_1016: TA to TA buffer manipulations
This test case is derived from gp_8040 which was found to be failing on
HiKey with a preliminary version of the 64-bit pager enabled
(CFG_ARM64_core=y CFG_WITH_PAGER=y CFG_PAGED_USER_TA=y). This new
regression_1016 test fails in the same circumstances, but also when
paging is enabled in a 32-bit build of the TEE core prior to commit
commit ecb06119e67d ("core: pager: invalidate tlb when clearing
entry").

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>