regression: 400x: test all subcases regardless of return value from TEE Cleint API.
In subcases of regression 40xx, if TEE Client API returns except TEEC_SUCCESS, xtest jumps to TEEC_CloseSession immediately.
So, the rest subcases are not tested.

This problem occurs, for example, when the parameters in the subcase are not supported by the cryptographic library in OPTEE OS.
In this case, TEE Client API returns TEEC_ERROR_NOT_SUPPORTED or TEEC_ERROR_BAD_PARAMETERS and the xtest closes the session leaving the rest subcases untested.

To avoid this, the xtest should jump to Do_ADBG_EndSubCase in the loop of subcases if TEE Client API returns except TEEC_SUCCESS.
This makes it possible to test all subcases regardless of the return code from TEE Client API.

Signed-off-by: Akira Shibakawa <akira.shibakawa@miraclelinux.com>