regression 6018: reduce even more blocks
We are still getting out-of-memory errors in 6018 when we have enabled
the pager and are using RPMB on HiKey. By reducing the number of blocks
written to RPMB a bit more we get rid of the problem. The problem seems
to be that the heap gets too fragmented when we are using the pager, so
even though this solves the problem here and now, we would need to look
into whether we can do something about the fragmented heap in the long
run.

Related, we have reduced this in the past:
  20949ea945f2 ("regression 6018: reduce the number of blocks written")

Fixes: https://github.com/OP-TEE/optee_os/issues/2580

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Tested-by: Joakim Bech <joakim.bech@linaro.org> (HiKey 6220)