[Fixup!] [RFC] Add test case for dlopen()/dlsym()/dlclose() (1021)
Remove the test that tries to get a function pointer to the test
function "ta_entry_call_lib_dl()", which is in the main TA executable.
This doesn't work, at least on QEMUv8, because the symbol is in .symtab
and not in .dynsym. We are only using .dynsym to resolve symbols,
because that's all we needed to link shared libraries. So we can't find
all symbols.

This feature is not needed at the moment so let's just drop the test.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>