regression: add tests for panic events handling
Adds a new set of tests for checking the proper releasing of
the memory resources in case of a panic event. The tests are
covering use cases when connections to the panicked context are
established whether to non-secure sessions (CA) or
other secure sessions (TA).

Also are tested several scenarios which involve all supporter
instance types:
- Multiple Instance TAs
- Single Instance TAs
- Single Instance Keep Alive TAs

Signed-off-by: Ovidiu Mihalachi <ovidiu_mihalachi@mentor.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>