regression 1010: Clang: prevent unwanted optimization
Calling a NULL function pointer is undefined behavior in C, therefore
when the compiler identifies such a pattern it is allowed to do pretty
much anything. Clang simply does not take the "case 3" branch in
ta_entry_bad_mem_access().

In order to force the call, this patch uses a volatile function pointer
so that the compiler cannot assume anything about its value.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>