build/Makefile: fix os_test with dynamic linking
When CFG_TA_DYNLINK is enabled, os_test depends on os_test_lib
to be build beforehand. When the top-level Makefile is not used,
e.g. in the case of buildroot, the build-order of tests is not
guaranteed. Especially when the test are ebumerated by globbing,
which depends on locale settings, os_test might be listed before
os_test_lib. Presence of the lib is ensured by making the dependency
explicit in the os_test-Makefile and executing a proper submake.

Signed-off-by: Markus S. Wamser <markus.wamser@mixed-mode.de>
Tested-by: Jerome Forissier <jerome@forissier.org> (make, buildroot)
Acked-by: Jerome Forissier <jerome@forissier.org>