regression 1025: skip test if memref null capability not supported
Skip regression test 1025 when TEE context reports that memref null
capability is not supported by the OP-TEE ecosystem. This change
prevents xtest suite to fail on regression 1025 when one tests
the latest optee_os/optee_client/optee_test that supports this recent
capability with a mainline or older Linux kernel that does not yet
support this feature.

Fixes: https://github.com/OP-TEE/optee_test/issues/415
Reported-by: Masami Hiramatsu <masami.hiramatsu@linaro.org>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome@forissier.org>