xtest: Linux next restriction on regression 1027 & 1028
Regression tests 1027 and 1028 mandates client UUID management in
Linux kernel OP-TEE driver. The feature is not merged in the mainline
kernel (as of v5.7-rc5). This change helps one to run xtest with
such test cases skipped by using kernel restrictions.

With this change, kernel next is needed (KERNEL_MAJOR_NEXT) to run
regression 1027 & 1028 when kernel restriction is enabled.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>