xtest: SPMC Add basic test
Add a first FF-A test. It will test the basic functionality of the FF-A
messages.
The tests are added to a new test suite(ffa_spmc) and can be run with the:
xtest -t ffa_spmc command

The corresponding SPs can be found in the trusted-services repo.

Currently the TS debugfs driver is needed to communicate with the SPs. It
can be loaded by running the
/mnt/host/out/linux-arm-ffa-user/load_module.sh script which is created
as part of the linux-arm-ffa-user target.

Signed-off-by: Jelle Sels <jelle.sels@arm.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>