xtest: Add test of gcov
Adds test for gcov in the range 31000

This patch adds a set of test of the generation of code coverage:
(At the font of the test suite)
 - 31001: Check the code coverage is enabled for the core:
 - 31002: Check the code coverage is enabled for the TA:
 - 31003: Dump the current code coverage of the core
 - 31004: Reset the code coverage of the core
(Normal execution in order, should be at the end)
 - 31005: Dump different code coverage:
   - 31005.1: Core code coverage from xtest
   - 31005.2: Core code coverage from TA
   - 31005.3: TA code coverage from TA

The test are added using ADBG_CASE_DEFINE_FRONT to the test suite
so the code coverage of the core is resetted before executing xtest
suite and dump at the end.
It allows to capture the code coverage of the core when running
xtest.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>