regression 1008: test open session failure once, unless level > 0
Test 1008 runs a negative test for TEEC_OpenSession() 100 times in a
loop, which can be useful to detect and troubleshoot memory leaks.
However the loop increases the execution time and the amount of logs
produced, so it is better reserved for level > 0 (xtest -l level).
Note, xtest 1013 reduces the number of loops when level == 0 already.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>