regression: don't use files in plugin tests
QEMU, xtest run while logged in as user test:

o regression_1033.2 Pass array to a plugin
regression_1000.c:2420: fp has an unexpected value: (nil), expected not NULL
  regression_1033.2 FAILED

The same test passes when logged in as root. The problem is, xtest tries to open,
read and delete a file created by tee-supplicant and permissions do not allow this.

Now request/response is used to test the ability to pass arrays to a plugin.
This approach is simpler and more robust.

Signed-off-by: Aleksandr Anisimov <a.anisimov@omprussia.ru>
Reviewed-by: Jerome Forissier <jerome@forissier.org>