xtest: pkcs11: Correct template for secret key
When creating object of type CKO_GENERIC_SECRET using
C_CreateObject(), CKA_VALUE is mandatory attribute in the
template and CKA_VALUE_LENGTH should not be provided. This
check was added in the implementation recently so the template
need to be corrected.

Also add an additional check to confirm that CKA_VALUE_LENGTH
gets populated correctly by the TA when such objects are created.

Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>