storage: add test cases with zero-sized object IDs
The TEE Internal Core API specification (v1.3.1) explicitly allows the
use of an empty object ID in TEE_RenamePersistentObject(). The text is:

 newObjectID, newObjectIDLen: A buffer containing the new object
   identifier. The identifier contains arbitrary bytes, including the
   zero byte. The identifier length SHALL be less than or equal to
   TEE_OBJECT_ID_MAX_LEN and can be zero.

(note the mention: "and can be zero").

Therefore, add a couple of test cases in regression_6008. Both
(non-NULL, 0) and (NULL, 0) are tested for the parameters
(newObjectID, newObjectIDLen).

Since it follows from the above that zero-sized object IDs are indeed
valid, similar test cases are added to regression_6001
(TEE_CreatePersistentObject()) and regression_6002
(TEE_OpenPersistentObject()).

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>