regression 4016: fix TEE_ATTR_EDDSA_PREHASH parameter
Prior to this patch was the TEE_ATTR_EDDSA_PREHASH attribute initialized
incorrectly. To indicate Ed25519ph the a and b values should be 1, 0
respectively. Giving the values 0, 0 is the same as not supplying the
TEE_ATTR_EDDSA_PREHASH attribute at all. So fix this by setting the
values 1, 0 when supplied.

Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>