xtest: sdp: remove ION support
The ION allocator is deprecated and not present in recent kernels.
Remove the code that uses it.

Tested on QEMUv8 with CFG_SECURE_DATA_PATH=n. Enabling SDP will
require a bit more work (use a toolchain with kernel headers recent
enough to have <linux/dma-heap.h>, and probably a few more things to
enable the SDP heap in the kernel), that is out of scope for this
commit.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>